cuisine fixes

+++ lib/JumpScale/tools/cuisine/apps/CuisineMongodb.py
@@ -8,7 +8,12 @@ app = j.tools.cuisine._getBaseAppClass()
 class CuisineMongodb(app):
     NAME = 'mongod'
 
-    def _build(self, reset=False):
+    def install(self, start=True, reset=False):
+        """
+        download, install, move files to appropriate places, and create relavent configs
+        """
+        if not reset and self.doneGet("install"):
+            return
         if reset and self.cuisine.core.isMac:
             self.cuisine.core.run("brew uninstall mongodb", die=False)
         if not reset and self.isInstalled():
@@ -20,7 +25,7 @@ class CuisineMongodb(app):
 
             url = None
             if self.cuisine.core.isUbuntu:
-                # TODO: *2 upgrade ubuntu
+                # TODO: *2 upgrade ubuntu version as well
                 # https://fastdl.mongodb.org/linux/mongodb-linux-x86_64-ubuntu1604-3.4.0.tgz
                 url = 'https://fastdl.mongodb.org/linux/mongodb-linux-x86_64-ubuntu1604-3.2.9.tgz'
             elif self.cuisine.core.isArch:
@@ -33,31 +38,31 @@ class CuisineMongodb(app):
             if url:
                 self.log('Downloading mongodb.')
                 self.cuisine.core.file_download(url, to="$TMPDIR", overwrite=False, expand=True)
-                tarpath = self.cuisine.core.find("$TMPDIR", recursive=True, pattern="*mongodb*.tgz", type='f')[0]
+                tarpaths = self.cuisine.core.find("$TMPDIR", recursive=False, pattern="*mongodb*.tgz", type='f')
+                if len(tarpaths) == 0:
+                    raise j.exceptions.Input(message="could not download:%s, did not find in %s" % (
+                        url, self.replace("$TMPDIR")), level=1, source="", tags="", msgpub="")
+                tarpath = tarpaths[0]
                 self.cuisine.core.file_expand(tarpath, "$TMPDIR")
                 extracted = self.cuisine.core.find("$TMPDIR", recursive=True, pattern="*mongodb*", type='d')[0]
                 for file in self.cuisine.core.find('%s/bin/' % extracted, type='f'):
                     self.cuisine.core.file_copy(file, appbase)
-
-    def install(self, start=True, name='mongod'):
-        """
-        download, install, move files to appropriate places, and create relavent configs
-        """
-        self.cuisine.core.dir_ensure('$VARDIR/data/%s' % name)
+        self.cuisine.core.dir_ensure('$VARDIR/data/mongodb')
+        self.doneSet("install")
         if start:
-            self.start(name)
+            self.start(reset=reset)
 
-    def build(self, start=True, install=True, reset=False):
-        self._build(reset=reset)
-        if install:
-            self.install(start)
+    def build(self, start=True, reset=False):
+        raise RuntimeError("not implemented")
 
-    def start(self, name="mongod"):
-        which = self.cuisine.core.command_location("mongod")
-        self.cuisine.core.dir_ensure('$VARDIR/data/%s' % name)
-        cmd = "%s --dbpath $VARDIR/data/%s" % (which, name)
-        # self.cuisine.process.kill("mongod")
-        self.cuisine.processmanager.ensure(name, cmd=cmd, env={}, path="")
+    def start(self, reset=False):
+        if not reset and self.doneGet("start"):
+            return
+        self.cuisine.core.dir_ensure('$VARDIR/data/mongodb')
+        cmd = "$BINDIR/mongod --dbpath '$VARDIR/data/mongodb'"
+        self.cuisine.process.kill("mongod")
+        self.cuisine.processmanager.ensure(name="mongod", cmd=cmd, env={}, path="")
+        self.doneSet("start")
 
-    def stop(self, name='mongod'):
-        self.cuisine.processmanager.stop(name)
+    def stop(self):
+        self.cuisine.processmanager.stop()
